+++ public/controllers/management/groups.js
@@ -1,5 +1,5 @@
 /*
- * Wazuh app - Groups controller
+ * Wazuh app - Management groups controller
  * Copyright (C) 2018 Wazuh, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -9,9 +9,9 @@
  *
  * Find more information about this on the LICENSE file.
  */
-import beautifier     from '../utils/json-beautifier';
+import beautifier     from '../../utils/json-beautifier';
 import { uiModules }  from 'ui/modules'
-import * as FileSaver from '../services/file-saver'
+import * as FileSaver from '../../services/file-saver'
 
 const app = uiModules.get('app/wazuh', []);
