+++ public/controllers/management/monitoring.js
@@ -9,7 +9,7 @@
  *
  * Find more information about this on the LICENSE file.
  */
-import FilterHandler from '../utils/filter-handler'
+import FilterHandler from '../../utils/filter-handler'
 import { uiModules } from 'ui/modules'
 
 const app = uiModules.get('app/wazuh', []);
