+++ public/controllers/misc/blank-screen-controller.js
