+++ public/controllers/misc/health-check.js
