+++ public/controllers/misc/reporting.js
