+++ public/controllers/settings/index.js
@@ -9,9 +9,9 @@
  *
  * Find more information about this on the LICENSE file.
  */
-import base64        from '../utils/base64.js';
+import base64        from '../../utils/base64';
 import { uiModules } from 'ui/modules';
-import TabNames      from '../utils/tab-names';
+import TabNames      from '../../utils/tab-names';
 
 const app = uiModules.get('app/wazuh', []);
