+++ public/templates/management/management.jade
@@ -1,0 +1,10 @@
+include ./management.head
+include ./welcome.html
+include ./status.html
+include ./configuration.html
+include ./logs.html
+include ./reporting.html
+include ./monitoring/monitoring.jade
+include ./groups/groups.html
+include ./ruleset/ruleset.jade
+include ../footer.foot
