+++ public/templates/management/monitoring/monitoring.jade
