+++ public/templates/management/monitoring/nodes-list.html
