+++ public/templates/management/reporting.html
