+++ public/templates/management/ruleset/decoders/decoders-list.html
