+++ public/templates/management/ruleset/rules/rules-list.html
