+++ public/templates/management/status.html
