--- public/templates/manager/manager.jade
@@ -1,10 +1,0 @@
-include ./manager.head
-include ./manager-welcome.html
-include ./manager-status.html
-include ./manager-configuration.html
-include ./manager-osseclog.html
-include ./manager-reporting.html
-include ./monitoring/monitoring.jade
-include ./groups/groups.html
-include ./ruleset/ruleset.jade
-include ../footer.foot
